chore: consolidate and tidy the test suite
F-14: AnnotationSuppressionLoggingTest asserted no suppression message at
INFO while the recipe logs at DEBUG, so the test could never fail. Assert
at DEBUG, correct the misleading comments, and drop the hardcoded
developer path from the class javadoc.

F-15: Delete the scaffolding/duplicate test classes (CuiLoggerRenameTest,
CuiLoggerRenameSimpleTest, CuiLoggerAutoFixTest,
CuiLogRecordPatternRecipeReproduceTest, InvalidExceptionUsageRecipeIssue5Test),
folding their unique scenarios into the main recipe test classes with a
javadoc issue reference. Deduplicate identical-fixture tests across
RecipeSuppressionUtilTest, AnnotationSuppressionTest and
BaseSuppressionVisitorTest.

F-16: Add coverage for the PathExclusionVisitor custom-pattern constructor
and its non-SourceFile branch, a recipe-metadata smoke test over all four
recipes, and a shared CuiLogger stub fixture used by the logging tests.

F-17: Parameterize the repetitive LogLevelTest, LombokAnnotationFormatTest
and AnnotationNewlineFormat cases; replace the undeclared AssertJ usage in
RecipeMarkerUtilTest with JUnit assertions; drop the coverage-only
reflection test and the unimplemented cui-rewrite:enable fixture; rename
the NormalizeFormatRecipe test scaffolding to TestNormalizeFormatRecipe.

Test-only change; no production sources touched.
